% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_draws.R
\name{compute_lambda}
\alias{compute_lambda}
\title{Compute the overall infectivity}
\usage{
compute_lambda(incid, w)
}
\arguments{
\item{incid}{a multidimensional array containing values of the incidence
for each time step (1st dimension), location (2nd dimension) and
pathogen/strain/variant (3rd dimension)}

\item{w}{a matrix with two columns, each containing the probability mass
function for the discrete serial interval for each of the two
pathogen/strain/variants, starting with the probability mass function
for day 0 in the first row, which should be 0. each column in the matrix
should sum to 1}
}
\value{
a multidimensional array containing values of the overall
infectiousness for each time step (1st dimension), location (2nd dimension)
and pathogen/strain/variant (3rd dimension). The overall infectiousness for
a given location and pathogen/strain/variant represents the sum of
the incidence for that location and that pathogen/strain/variant at all
previous time steps, weighted by the current infectiousness of those
past incident cases.
}
\description{
Compute the overall infectivity
}
