% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_draws.R
\name{default_mcmc_controls}
\alias{default_mcmc_controls}
\title{Set default for MCMC controls}
\usage{
default_mcmc_controls()
}
\value{
a list of default MCMC control parameters, containing:
\itemize{
\item n_iter: the number if iterations of the MCMC to perform
\item burnin: the burnin to use; MCMC iterations will only be recorded after
the burnin
\item thin: MCMC iterations will only be recorded after
the burnin and every \code{thin} iteration

Values can then be manually be edited as in the examples below.
}
}
\description{
Set default for MCMC controls
}
\examples{
mcmc_controls <- default_mcmc_controls()
# change to run for 10 times longer
mcmc_controls$n_iter <- mcmc_controls$n_iter * 10

}
