% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbs_draws.R
\name{estimate_joint}
\alias{estimate_joint}
\title{Jointly estimate the instantaneous reproduction number for a reference
pathogen/strain/variant and the relative transmissibility of a
"new" pathogen/strain/variant}
\usage{
estimate_joint(
  epsilon_init,
  R_init,
  incid,
  w,
  priors,
  n_iter = 1000,
  burnin = 10,
  t_min = 2,
  t_max = nrow(incid),
  seed = NULL
)
}
\arguments{
\item{epsilon_init}{value at which epsilon will be initialised. Epsilon
is the relative transmissibility of the "new"
pathogen/strain/variant compared to the reference pathogen/strain/variant}

\item{R_init}{a matrix with dimensions containing the initial values of the
instantaneous reproduction number for each time step (row) and location
(column), for the reference pathogen/strain/variant}

\item{incid}{a multidimensional array containing values of the incidence
for each time step (1st dimension), location (2nd dimension) and
pathogen/strain/variant (3rd dimension)}

\item{w}{a matrix with two columns, each containing the probability mass
function for the discrete serial interval for each of the two
pathogen/strain/variants, starting with the probability mass function
for day 0 in the first row, which should be 0. each column in the matrix
should sum to 1}

\item{priors}{a list of prior parameters (shape and scale of a gamma
distribution) for epsilon and R; can be obtained from the function
\code{default_priors}. The prior for R is assumed to be the same for all
time steps and all locations}

\item{n_iter}{the number if iterations of the MCMC to perform}

\item{burnin}{the burnin to use; MCMC iterations will only be recorded after
the burnin}

\item{t_min}{an integer >1 giving the minimum time step to consider in the
estimation. Default value is 2 (as the estimation is conditional on
observations at time step 1 and can therefore only start at time step 2).}

\item{t_max}{an integer >\code{t_min} and <=\code{nrow(incid)} giving the maximum time
step to consider in the estimation. Default value is \code{nrow(incid)}.}

\item{seed}{a numeric value used to fix the random seed}
}
\value{

}
\description{
Jointly estimate the instantaneous reproduction number for a reference
pathogen/strain/variant and the relative transmissibility of a
"new" pathogen/strain/variant
}
\examples{

n_loc <- 3 # 3 locations
T <- 100 # 100 time steps
priors <- default_priors()
# constant incidence 10 per day everywhere
incid <- array(10, dim = c(T, n_loc, n_v))
# arbitrary serial interval, same for both variants
w_v <- c(0, 0.2, 0.5, 0.3)
w <- cbind(w_v, w_v)

# Dummy initial values for the MCMC
R_init <- matrix(5, nrow = T, ncol = n_loc)
R_init[1, ] <- NA # no estimates of R on first time step
epsilon_init <- 5
x <- estimate_joint(epsilon_init, R_init,
                    incid, w, priors,
                    n_iter = 1000,
                    burnin = 10)
# Plotting to check outputs
par(mfrow = c(2, 2))
plot(x$epsilon, type = "l",
     xlab = "Iteration", ylab = "epsilon")
# Compare with what we expect with constant incidence in all locations
abline(h = 1, col = "red")
plot(x$R[10, 1, ], type = "l",
     xlab = "Iteration", ylab = "R time 10 location 1")
abline(h = 1, col = "red")
plot(x$R[20, 2, ], type = "l",
     xlab = "Iteration", ylab = "R time 20 location 2")
abline(h = 1, col = "red")
plot(x$R[30, 3, ], type = "l",
     xlab = "Iteration", ylab = "R time 30 location 3")

}
